+++ packages/better_auth/test/better_auth/plugins/external_plugin_shim_test.rb
@@ -4,47 +4,42 @@ require_relative "../../test_helper"
 
 class ExternalPluginShimTest < Minitest::Test
   def test_sso_shim_has_helpful_error_when_external_package_is_missing
-    error = assert_raises(LoadError) do
-      BetterAuth::Plugins.sso
-    end
-
-    assert_includes error.message, "better_auth-sso"
-    assert_includes error.message, "require \"better_auth/sso\""
+    assert_external_plugin_shim(:sso, "better_auth-sso", "better_auth/sso")
   end
 
   def test_scim_shim_has_helpful_error_when_external_package_is_missing
-    error = assert_raises(LoadError) do
-      BetterAuth::Plugins.scim
-    end
-
-    assert_includes error.message, "better_auth-scim"
-    assert_includes error.message, "require \"better_auth/scim\""
+    assert_external_plugin_shim(:scim, "better_auth-scim", "better_auth/scim")
   end
 
   def test_passkey_shim_has_helpful_error_when_external_package_is_missing
-    error = assert_raises(LoadError) do
-      BetterAuth::Plugins.passkey
-    end
-
-    assert_includes error.message, "better_auth-passkey"
-    assert_includes error.message, "require \"better_auth/passkey\""
+    assert_external_plugin_shim(:passkey, "better_auth-passkey", "better_auth/passkey")
   end
 
   def test_oauth_provider_shim_has_helpful_error_when_external_package_is_missing
-    error = assert_raises(LoadError) do
-      BetterAuth::Plugins.oauth_provider
-    end
-
-    assert_includes error.message, "better_auth-oauth-provider"
-    assert_includes error.message, "require \"better_auth/oauth_provider\""
+    assert_external_plugin_shim(:oauth_provider, "better_auth-oauth-provider", "better_auth/oauth_provider")
   end
 
   def test_api_key_shim_has_helpful_error_when_external_package_is_missing
-    error = assert_raises(LoadError) do
-      BetterAuth::Plugins.api_key
+    assert_external_plugin_shim(:api_key, "better_auth-api-key", "better_auth/api_key")
+  end
+
+  private
+
+  def assert_external_plugin_shim(method_name, gem_name, require_path)
+    original_require = Kernel.method(:require)
+    missing_require = lambda do |path|
+      raise LoadError, "cannot load such file -- #{path}" if path == require_path
+
+      original_require.call(path)
     end
 
-    assert_includes error.message, "better_auth-api-key"
-    assert_includes error.message, "require \"better_auth/api_key\""
+    error = assert_raises(LoadError) do
+      Kernel.stub(:require, missing_require) do
+        BetterAuth::Plugins.public_send(method_name)
+      end
+    end
+
+    assert_includes error.message, gem_name
+    assert_includes error.message, "require \"#{require_path}\""
   end
 end